添加后台 job 岗位管理

+++ module-job/module-job-model/src/main/java/io/jpress/module/job/model/base/BaseJobCategory.java
@@ -92,15 +92,15 @@ public abstract class BaseJobCategory<M extends BaseJobCategory<M>> extends Base
     /**
      * 该分类下的岗位数量
      */
-	public void setCount(java.lang.Integer count) {
+	public void setCount(java.lang.Long count) {
 		set("count", count);
 	}
 
     /**
      * 该分类下的岗位数量
      */
-	public java.lang.Integer getCount() {
-		return getInt("count");
+	public java.lang.Long getCount() {
+		return getLong("count");
 	}
 
     /**
